+++ src/chain/tezos/TezosNodeWriter.ts
@@ -374,7 +374,7 @@ export namespace TezosNodeWriter {
         amount: number,
         delegate: string | undefined,
         fee: number,
-        derivationPath: string,
+        derivationPath: string | undefined,
         storage_limit: number,
         gas_limit: number,
         code: string,
@@ -617,7 +617,10 @@ export namespace TezosNodeWriter {
             const arr = Array.isArray(json) ? json : [json];
 
             if ('kind' in arr[0]) {
+                // TODO: show counter errors better: e.msg "already used for contract"
                 errors = arr.map(e => `(${e.kind}: ${e.id})`).join('');
+            } else if (arr.length === 1 && arr[0].contents.length === 1 && arr[0].contents[0].kind === 'activate_account') {
+                // in CARTHAGE and prior protocols activation failures are caught in the first branch
             } else {
                 errors = arr.map(r => r.contents)
                         .map(o => o.map(c => c.metadata.operation_result)